Inline core.rewriter package into core.statement package.
