+++ core/src/main/java/org/jdbi/v3/core/statement/ColonPrefixSqlParser.java
@@ -11,16 +11,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.v3.core.rewriter;
+package org.jdbi.v3.core.statement;
 
-import static org.jdbi.v3.core.internal.lexer.HashStatementLexer.COMMENT;
-import static org.jdbi.v3.core.internal.lexer.HashStatementLexer.DOUBLE_QUOTED_TEXT;
-import static org.jdbi.v3.core.internal.lexer.HashStatementLexer.EOF;
-import static org.jdbi.v3.core.internal.lexer.HashStatementLexer.ESCAPED_TEXT;
-import static org.jdbi.v3.core.internal.lexer.HashStatementLexer.LITERAL;
-import static org.jdbi.v3.core.internal.lexer.HashStatementLexer.NAMED_PARAM;
-import static org.jdbi.v3.core.internal.lexer.HashStatementLexer.POSITIONAL_PARAM;
-import static org.jdbi.v3.core.internal.lexer.HashStatementLexer.QUOTED_TEXT;
+import static org.jdbi.v3.core.internal.lexer.ColonStatementLexer.*; // SUPPRESS CHECKSTYLE WARNING
 
 import java.util.Collections;
 import java.util.Map;
@@ -28,15 +21,16 @@ import java.util.WeakHashMap;
 
 import org.antlr.runtime.ANTLRStringStream;
 import org.antlr.runtime.Token;
-import org.jdbi.v3.core.internal.lexer.HashStatementLexer;
-import org.jdbi.v3.core.statement.StatementContext;
-import org.jdbi.v3.core.statement.UnableToCreateStatementException;
+import org.jdbi.v3.core.internal.lexer.ColonStatementLexer;
 
 /**
  * SQL parser which recognizes named parameter tokens of the form
- * <code>#tokenName</code>.
+ * <code>:tokenName</code>
+ * <p>
+ * This is the default SQL parser
+ * </p>
  */
-public class HashPrefixSqlParser implements SqlParser {
+public class ColonPrefixSqlParser implements SqlParser {
     private final Map<String, ParsedSql> cache = Collections.synchronizedMap(new WeakHashMap<>());
 
     @Override
@@ -48,9 +42,9 @@ public class HashPrefixSqlParser implements SqlParser {
         }
     }
 
-    private ParsedSql internalParse(final String sql) {
+    private ParsedSql internalParse(String sql) throws IllegalArgumentException {
         ParsedSql.Builder parsedSql = ParsedSql.builder();
-        HashStatementLexer lexer = new HashStatementLexer(new ANTLRStringStream(sql));
+        ColonStatementLexer lexer = new ColonStatementLexer(new ANTLRStringStream(sql));
         Token t = lexer.nextToken();
         while (t.getType() != EOF) {
             switch (t.getType()) {
