+++ core/src/main/java/org/jdbi/v3/core/statement/ParsedParameters.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.v3.core.rewriter;
+package org.jdbi.v3.core.statement;
 
 import static java.util.Collections.unmodifiableList;
