+++ core/src/main/java/org/jdbi/v3/core/statement/TemplateEngine.java
@@ -11,9 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.v3.core.rewriter;
-
-import org.jdbi.v3.core.statement.StatementContext;
+package org.jdbi.v3.core.statement;
 
 /**
  * Renders an SQL statement from a template.
