+++ core/src/test/java/org/jdbi/v3/core/statement/TestEscapedCharacters.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.v3.core.rewriter;
+package org.jdbi.v3.core.statement;
 
 import org.jdbi.v3.core.statement.StatementContext;
 import org.junit.Test;
