+++ core/src/test/java/org/jdbi/v3/core/statement/TestHashPrefixSqlParser.java
@@ -11,24 +11,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.v3.core.rewriter;
+package org.jdbi.v3.core.statement;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.when;
 
 import java.util.Collections;
 import java.util.Map;
 
+import com.google.common.collect.ImmutableMap;
+
 import org.jdbi.v3.core.statement.StatementContext;
+import org.jdbi.v3.core.statement.StatementContextAccess;
 import org.jdbi.v3.core.statement.UnableToCreateStatementException;
 import org.junit.Before;
 import org.junit.Test;
 
-import com.google.common.collect.ImmutableMap;
-
-public class TestColonPrefixTemplateEngine
+public class TestHashPrefixSqlParser
 {
     private TemplateEngine templateEngine;
     private SqlParser parser;
@@ -37,8 +36,8 @@ public class TestColonPrefixTemplateEngine
     @Before
     public void setUp() throws Exception
     {
-        templateEngine = new DefinedAttributeTemplateEngine();
-        parser = new ColonPrefixSqlParser();
+        this.templateEngine = new DefinedAttributeTemplateEngine();
+        this.parser = new HashPrefixSqlParser();
         ctx = mock(StatementContext.class);
     }
 
@@ -48,7 +47,8 @@ public class TestColonPrefixTemplateEngine
     }
 
     private String render(String sql, Map<String, Object> attributes) {
-        attributes.forEach((key, value) -> when(ctx.getAttribute(key)).thenReturn(value));
+        StatementContext ctx = StatementContextAccess.createContext();
+        attributes.forEach(ctx::define);
 
         return templateEngine.render(sql, ctx);
     }
@@ -56,22 +56,22 @@ public class TestColonPrefixTemplateEngine
     @Test
     public void testNewlinesOkay() throws Exception
     {
-        ParsedSql parsed = parser.parse("select * from something\n where id = :id", ctx);
+        ParsedSql parsed = parser.parse("select * from something\n where id = #id", ctx);
         assertThat(parsed.getSql()).isEqualTo("select * from something\n where id = ?");
     }
 
     @Test
     public void testOddCharacters() throws Exception
     {
-        ParsedSql parsed = parser.parse("~* :boo ':nope' _%&^& *@ :id", ctx);
-        assertThat(parsed.getSql()).isEqualTo("~* ? ':nope' _%&^& *@ ?");
+        ParsedSql parsed = parser.parse("~* #boo '#nope' _%&^& *@ #id", ctx);
+        assertThat(parsed.getSql()).isEqualTo("~* ? '#nope' _%&^& *@ ?");
     }
 
     @Test
     public void testNumbers() throws Exception
     {
-        ParsedSql parsed = parser.parse(":bo0 ':nope' _%&^& *@ :id", ctx);
-        assertThat(parsed.getSql()).isEqualTo("? ':nope' _%&^& *@ ?");
+        ParsedSql parsed = parser.parse("#bo0 '#nope' _%&^& *@ #id", ctx);
+        assertThat(parsed.getSql()).isEqualTo("? '#nope' _%&^& *@ ?");
     }
 
     @Test
@@ -82,10 +82,10 @@ public class TestColonPrefixTemplateEngine
     }
 
     @Test
-    public void testHashInColumnNameOkay() throws Exception
+    public void testColonIsLiteral() throws Exception
     {
-        ParsedSql parsed = parser.parse("select column# from thetable where id = :id", ctx);
-       assertThat(parsed.getSql()).isEqualTo("select column# from thetable where id = ?");
+        ParsedSql parsed = parser.parse("select * from foo where id = :id", ctx);
+        assertThat(parsed.getSql()).isEqualTo("select * from foo where id = :id");
     }
 
     @Test
@@ -95,18 +95,10 @@ public class TestColonPrefixTemplateEngine
         assertThat(parsed.getSql()).isEqualTo("select * from `v$session");
     }
 
-    @Test
-    public void testDoubleColon() throws Exception
-    {
-        final String doubleColon = "select 1::int";
-        ParsedSql parsed = parser.parse(doubleColon, ctx);
-        assertThat(parsed.getSql()).isEqualTo(doubleColon);
-    }
-
     @Test(expected = UnableToCreateStatementException.class)
     public void testBailsOutOnInvalidInput() throws Exception
     {
-        render("select * from something\n where id = :\u0087\u008e\u0092\u0097\u009c");
+        parser.parse("select * from something\n where id = #\u0087\u008e\u0092\u0097\u009c", ctx);
     }
 
     @Test
@@ -115,7 +107,7 @@ public class TestColonPrefixTemplateEngine
         Map<String, Object> attributes = ImmutableMap.of(
                 "column", "foo",
                 "table", "bar");
-        String rendered = render("select <column> from <table> where <column> = :someValue", attributes);
+        String rendered = render("select <column> from <table> where <column> = #someValue", attributes);
         ParsedSql parsed = parser.parse(rendered, ctx);
         assertThat(parsed.getSql()).isEqualTo("select foo from bar where foo = ?");
     }
@@ -141,25 +133,9 @@ public class TestColonPrefixTemplateEngine
     }
 
     @Test
-    public void testCachesRewrittenStatements() throws Exception
-    {
-        parser = spy(parser);
-
-        String sql = "insert into something (id, name) values (:id, :name)";
-        ParsedSql parsed = parser.parse(sql, ctx);
-        assertThat(parsed).isSameAs(parser.parse(sql, ctx));
-    }
-
-    @Test
     public void testCommentQuote() throws Exception
     {
         String sql = "select 1 /* ' \" <foo> */";
-        assertThat(render(sql)).isEqualTo(sql);
-    }
-
-    @Test
-    public void testColonInComment() throws Exception {
-        String sql = "/* comment with : colons :: inside it */ select 1";
-        assertThat(render(sql)).isEqualTo(sql);
+        assertThat(parser.parse(sql, ctx).getSql()).isEqualTo(sql);
     }
 }
