+++ sqlobject/src/main/java/org/jdbi/v3/sqlobject/config/UseSqlParser.java
@@ -18,7 +18,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-import org.jdbi.v3.core.rewriter.SqlParser;
+import org.jdbi.v3.core.statement.SqlParser;
 import org.jdbi.v3.sqlobject.config.internal.UseSqlParserImpl;
 
 /**
