+++ stringtemplate4/src/main/java/org/jdbi/v3/stringtemplate4/StringTemplateEngine.java
@@ -13,7 +13,7 @@
  */
 package org.jdbi.v3.stringtemplate4;
 
-import org.jdbi.v3.core.rewriter.TemplateEngine;
+import org.jdbi.v3.core.statement.TemplateEngine;
 import org.jdbi.v3.core.statement.StatementContext;
 import org.stringtemplate.v4.ST;
