+++ stringtemplate4/src/main/java/org/jdbi/v3/stringtemplate4/internal/UseStringTemplateSqlLocatorImpl.java
@@ -18,7 +18,7 @@ import static org.jdbi.v3.stringtemplate4.StringTemplateSqlLocator.findStringTem
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
 import org.jdbi.v3.core.config.ConfigRegistry;
-import org.jdbi.v3.core.rewriter.TemplateEngine;
+import org.jdbi.v3.core.statement.TemplateEngine;
 import org.jdbi.v3.core.statement.SqlStatements;
 import org.jdbi.v3.sqlobject.SqlObjects;
 import org.jdbi.v3.sqlobject.config.Configurer;
